修改预警类

+++ src/main/java/com/hbhb/cw/relocation/model/RelocationWarn.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -22,11 +24,11 @@ public class RelocationWarn {
 
     private String contractNum;
 
-    private Long anticipatePayment;
+    private BigDecimal anticipatePayment;
 
     private Byte isReceived;
 
-    private Long finalPayment;
+    private BigDecimal finalPayment;
 
     private Integer contractDuration;
